Add files via upload
Test/Train files with missing values replaced with mean/median/dropped values. Mean/Median in each file is used with respect to itself. For example, mean of test is used in test and the mean of the training set is used in train. I think it's probably proper procedure to use the mean  of the training set for missing values in both test and train... But I am not certain. Categorical variables are converted to dummies so they should be good to go.